+++ lib/components/dropdown_menus.dart
@@ -3,126 +3,73 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_waya/flutter_waya.dart';
 
-typedef DropdownMenusTitleBuilder = Widget Function(
-    BuildContext context, int index, bool visible);
+typedef DropdownMenusChanged<K, V> = void Function(K key, V? value);
 
-typedef DropdownMenusIconBuilder = Widget Function(bool selected);
-
-typedef DropdownMenusValueCallback = void Function(
-    int labelIndex, int? valueIndex);
-
-typedef DropdownMenusValueBuilder = Widget Function(
-    BuildContext context, int labelIndex, int valueIndex);
-
-class DropdownMenusItem<T> {
-  const DropdownMenusItem({required this.label, this.onTap});
+class DropdownMenusItem {
+  const DropdownMenusItem({required this.child, this.onTap});
 
   /// label
-  final Widget label;
+  final Widget child;
 
   /// onTap
-  final ValueCallback<T>? onTap;
+  final GestureTapCallback? onTap;
 }
 
-class DropdownMenusLabelItem<T, K> extends DropdownMenusItem<T> {
-  DropdownMenusLabelItem(
-      {required super.label,
-      required this.items,
-      this.icon,
-      super.onTap,
-      this.isRotate = true});
+class DropdownMenusValueItem<V> extends DropdownMenusItem {
+  const DropdownMenusValueItem(
+      {required super.child, super.onTap, required this.value});
 
-  /// icon
+  /// value
+  final V value;
+}
+
+class DropdownMenusKeyItem<K, V> extends DropdownMenusItem {
+  DropdownMenusKeyItem({
+    required super.child,
+    required this.value,
+    super.onTap,
+    this.items = const [],
+    this.icon,
+    this.enabled = true,
+  });
+
+  /// icon 自动旋转
   final Widget? icon;
 
-  /// icon 是否 添加旋转
-  final bool isRotate;
-
   /// items
-  final List<DropdownMenusItem<K>> items;
+  final List<DropdownMenusValueItem<V>> items;
+
+  /// value
+  final K value;
+
+  /// 是否可点击
+  final bool enabled;
+
+  bool rotateState = false;
+
+  set setRotateState(bool value) {
+    rotateState = value;
+  }
 }
 
 /// 下拉菜单
-class DropdownMenus extends StatefulWidget {
-  DropdownMenus({
-    super.key,
-    TextStyle? labelStyle,
-    TextStyle? valueStyle,
-    this.backgroundColor = const Color(0x80000000),
-    required List<String> label,
-    required List<List<String>> value,
-    this.width,
-    this.decoration,
-    DropdownMenusIconBuilder? icon,
-    this.mainAxisAlignment = MainAxisAlignment.spaceAround,
-    this.margin,
-    this.padding = const EdgeInsets.symmetric(vertical: 10),
-    this.isModal = false,
-    this.onTap,
-    this.isRotate = true,
-  })  : assert(label.isNotEmpty, 'label cannot be empty'),
-        assert(label.length == value.length,
-            'the length of label and value must be consistent'),
-        labelCount = label.length,
-        valueCount = value.builder((List<String> item) => item.length),
-        icon = icon ??
-            ((bool selected) => const Icon(Icons.keyboard_arrow_up, size: 20)),
-        labelBuilder = ((BuildContext context, int index, bool isSelect) =>
-            Padding(
-                padding: const EdgeInsets.only(right: 6),
-                child: BText(label[index],
-                    style: labelStyle ?? context.textTheme.labelSmall))),
-        valueBuilder = ((BuildContext context, int labelIndex,
-                int valueIndex) =>
-            Container(
-                alignment: Alignment.center,
-                width: double.infinity,
-                padding: const EdgeInsets.symmetric(vertical: 10),
-                decoration: BoxDecoration(
-                    color: context.theme.canvasColor,
-                    border: Border(
-                        top: BorderSide(color: context.theme.dividerColor))),
-                child: BText(value[labelIndex][valueIndex],
-                    style: valueStyle ?? context.textTheme.bodyLarge)));
+class DropdownMenus<K, V> extends StatefulWidget {
+  const DropdownMenus(
+      {super.key,
+      required this.menus,
+      this.backgroundColor = const Color(0x80000000),
+      this.width,
+      this.isModal = false,
+      this.decoration,
+      this.mainAxisAlignment = MainAxisAlignment.spaceAround,
+      this.margin,
+      this.padding,
+      this.onChanged});
 
-  const DropdownMenus.custom({
-    super.key,
-    required this.labelCount,
-    required this.labelBuilder,
-    required this.valueCount,
-    required this.valueBuilder,
-    this.icon,
-    this.width,
-    this.decoration,
-    this.backgroundColor = const Color(0x80000000),
-    this.mainAxisAlignment = MainAxisAlignment.spaceAround,
-    this.margin,
-    this.padding = const EdgeInsets.symmetric(vertical: 10),
-    this.isModal = false,
-    this.onTap,
-    this.isRotate = true,
-  }) : assert(labelCount == valueCount.length);
+  /// 子menus
+  final List<DropdownMenusKeyItem<K, V>> menus;
 
-  /// label 长度
-  final int labelCount;
-
-  /// label 每个item
-  final DropdownMenusTitleBuilder labelBuilder;
-
-  /// value 长度
-  final List<int> valueCount;
-
-  /// value 每个item
-  final DropdownMenusValueBuilder valueBuilder;
-
-  /// icon 是否 添加旋转
-  final bool isRotate;
-
-  /// 自定义icon
-  final DropdownMenusIconBuilder? icon;
-
-  /// 点击回调 valueIndex == null 表示 只是点击了label
-  final DropdownMenusValueCallback? onTap;
+  final DropdownMenusChanged<K, V>? onChanged;
 
   /// value 显示时 的背景色
   final Color backgroundColor;
@@ -133,59 +80,33 @@ class DropdownMenus extends StatefulWidget {
   /// 是否模态 [isModal]=true,背景点击无响应
   final bool isModal;
 
-  /// 以下属性作用于 label
   final Decoration? decoration;
   final MainAxisAlignment mainAxisAlignment;
   final EdgeInsetsGeometry? margin;
   final EdgeInsetsGeometry? padding;
 
   @override
-  State<DropdownMenus> createState() => _DropdownMenusState();
+  State<DropdownMenus<K, V>> createState() => _DropdownMenusState<K, V>();
 }
 
-class _DropdownMenusState extends State<DropdownMenus> {
-  List<bool> labelState = <bool>[];
+class _DropdownMenusState<K, V> extends State<DropdownMenus<K, V>> {
+  late List<DropdownMenusKeyItem<K, V>> menus;
+
   late GlobalKey labelKey = GlobalKey();
 
-  void changeState(int index) => setState(() {
-        labelState[index] = !labelState[index];
-      });
+  @override
+  void initState() {
+    menus = widget.menus;
+    super.initState();
+  }
 
-  void popupWidget(int index) {
-    final NavigatorState navigator = Navigator.of(context);
-    final Rect? rect = context.getWidgetRectLocalToGlobal(
-        ancestor: navigator.context.findRenderObject());
-    final ScrollList listBuilder = ScrollList.builder(
-        itemCount: widget.valueCount[index],
-        physics: const ClampingScrollPhysics(),
-        itemBuilder: (_, int i) => Universal(
-            alignment: Alignment.center,
-            width: double.infinity,
-            onTap: () {
-              changeState(index);
-              pop();
-              if (widget.onTap != null) widget.onTap!(index, i);
-            },
-            child: widget.valueBuilder(context, index, i)));
-    ModalWindows(
-            options: GlobalOptions().modalWindowsOptions.copyWith(
-                top: (rect?.top ?? 0) + context.size!.height,
-                alignment: Alignment.topLeft,
-                color: Colors.transparent,
-                onTap: widget.isModal
-                    ? null
-                    : () {
-                        changeState(index);
-                        pop();
-                      }),
-            child: Universal(
-                width: widget.width,
-                color: widget.backgroundColor,
-                child: listBuilder))
-        .popupDialog<dynamic>(
-            options: const DialogOptions(
-                fromStyle: PopupFromStyle.fromCenter,
-                barrierColor: Colors.transparent));
+  @override
+  void didUpdateWidget(covariant DropdownMenus<K, V> oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.menus != menus) {
+      menus = widget.menus;
+      setState(() {});
+    }
   }
 
   @override
@@ -198,35 +119,81 @@ class _DropdownMenusState extends State<DropdownMenus> {
         mainAxisAlignment: widget.mainAxisAlignment,
         direction: Axis.horizontal,
         decoration: widget.decoration,
-        children: widget.labelCount.generate((int index) {
-          labelState.add(false);
-          if (widget.icon != null && widget.isRotate) {
+        children: menus.builder((entry) {
+          if (entry.icon != null) {
             return ToggleRotate(
                 rad: pi,
-                isRotate: labelState[index],
-                toggleBuilder: (Widget child) =>
-                    Row(mainAxisSize: MainAxisSize.min, children: [
-                      widget.labelBuilder(context, index, labelState[index]),
-                      child
-                    ]),
-                child: widget.icon!(labelState[index]),
-                onTap: () => onTap(index));
+                isRotate: !entry.rotateState,
+                toggleBuilder: (Widget child) => Row(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [entry.child, child]),
+                onTap: () => labelTap(entry),
+                child: entry.icon!);
           } else {
-            return widget
-                .labelBuilder(context, index, labelState[index])
-                .onTap(() => onTap(index));
+            return entry.child.onTap(() => labelTap(entry));
           }
         }));
   }
 
-  Future<void> onTap(int index) async {
-    widget.onTap?.call(index, null);
+  Future<void> labelTap(DropdownMenusKeyItem<K, V> entry) async {
+    if (!entry.enabled) return;
+    widget.onChanged?.call(entry.value, null);
+    entry.onTap?.call();
+    if (entry.items.isEmpty) return;
     final double keyboardHeight = getViewInsets.bottom;
     if (keyboardHeight > 0) {
       context.unfocus();
-      await 300.milliseconds.delayed(() {});
+      await 200.milliseconds.delayed(() {});
     }
-    changeState(index);
-    popupWidget(index);
+    changeRotateState(entry);
+    popupItems(entry);
+  }
+
+  void changeRotateState(DropdownMenusKeyItem<K, V> entry) {
+    entry.rotateState = !entry.rotateState;
+    setState(() {});
+  }
+
+  void popupItems(DropdownMenusKeyItem<K, V> entry) {
+    final items = entry.items;
+    final NavigatorState navigator = Navigator.of(context);
+    final Rect? rect = context.getWidgetRectLocalToGlobal(
+        ancestor: navigator.context.findRenderObject());
+    final ScrollList listBuilder = ScrollList.builder(
+        itemCount: items.length,
+        physics: const ClampingScrollPhysics(),
+        itemBuilder: (_, int i) {
+          final item = items[i];
+          return Universal(
+              alignment: Alignment.center,
+              width: double.infinity,
+              onTap: () {
+                item.onTap?.call();
+                widget.onChanged?.call(entry.value, item.value);
+                changeRotateState(entry);
+                pop();
+              },
+              child: item.child);
+        });
+    ModalWindows(
+            options: GlobalOptions().modalWindowsOptions.copyWith(
+                top: (rect?.top ?? 0) + context.size!.height,
+                alignment: Alignment.topLeft,
+                color: Colors.transparent,
+                onTap: widget.isModal
+                    ? null
+                    : () {
+                        widget.onChanged?.call(entry.value, null);
+                        changeRotateState(entry);
+                        pop();
+                      }),
+            child: Universal(
+                width: widget.width,
+                color: widget.backgroundColor,
+                child: listBuilder))
+        .popupDialog<dynamic>(
+            options: const DialogOptions(
+                fromStyle: PopupFromStyle.fromCenter,
+                barrierColor: Colors.transparent));
   }
 }